requirements.txt
absl-py==0.8.1
alabaster==0.7.12
allennlp==0.9.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.7.0
blis==0.2.4
boto==2.49.0
boto3==1.10.5
botocore==1.13.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
conllu==1.3.1
cycler==0.10.0
cymem==2.0.2
deepsegment==2.2.2
docutils==0.15.2
editdistance==0.5.3
en-core-web-sm==2.1.0
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
ftfy==5.6
gast==0.2.2
gensim==3.8.1
gevent==1.4.0
google-auth==1.6.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
greenlet==0.4.15
grpcio==1.24.3
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
inexactsearch==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
jsonnet==0.14.0
jsonpickle==1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
language-check==1.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
more-itertools==7.2.0
murmurhash==1.0.2
nltk==3.4.5
numpy==1.17.2
numpydoc==0.9.1
oauthlib==3.1.0
opt-einsum==3.1.0
overrides==2.4
packaging==19.2
pandas==0.25.1
parsimonious==0.8.1
plac==0.9.6
pluggy==0.13.0
preshed==2.0.1
progressbar2==3.47.0
protobuf==3.10.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycontractions==2.0.1
pydload==1.0.6
pyemd==0.5.1
Pygments==2.4.2
pyparsing==2.4.2
pyspellchecker==0.5.2
pytest==5.2.2
python-dateutil==2.8.0
python-utils==2.3.0
pytorch-pretrained-bert==0.6.2
pytorch-transformers==1.1.0
pytz==2019.3
PyYAML==5.1.2
regex==2019.8.19
requests==2.22.0
requests-oauthlib==1.2.0
responses==0.10.6
rsa==4.0
s3transfer==0.2.1
scikit-learn==0.21.3
scipy==1.3.1
sentencepiece==0.1.83
seqeval==0.0.3
seqtag-keras==1.0.5
silpa-common==0.3
six==1.12.0
smart-open==1.8.4
snowballstemmer==2.0.0
soundex==1.1.3
spacy==2.1.9
spellchecker==0.4
Sphinx==2.2.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
srsly==0.2.0
tensorboard==2.0.1
tensorboardX==1.9
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
thinc==7.0.8
torch==1.3.0
tqdm==4.36.1
tweet-preprocessor==0.5.0
Unidecode==1.1.1
urllib3==1.25.6
wasabi==0.3.0
wcwidth==0.1.7
Werkzeug==0.16.0
word2number==1.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.2
zipp==0.6.0
